stuck now on formatting of payment results

+++ app.rb
@@ -52,7 +52,7 @@ end
 
 get("/payment/results") do
 @apr = params.fetch("user_apr").to_f
-@years = params.fetch("user_years").to_i
+@years = params.fetch("user_years").to_f
 @principal = params.fetch("user_pv").to_f
 
 period_rate = @apr / 1200
